@@ -26,18 +26,14 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding, CafeteriaViewMo
 
     private lateinit var stuKoreanMenuAdapter: CafeteriaAdapter
     private lateinit var stuAnotherMenuAdapter: CafeteriaAdapter
-    private lateinit var eduKoreanMenuAdapter: CafeteriaAdapter
-    private lateinit var eduAnotherMenuAdapter: CafeteriaAdapter
 
     override fun initStartView() {
         binding.vm = viewModel
 
         stuKoreanMenuAdapter = CafeteriaAdapter()
         stuAnotherMenuAdapter = CafeteriaAdapter()
-        eduKoreanMenuAdapter = CafeteriaAdapter()
-        eduAnotherMenuAdapter = CafeteriaAdapter()
 
-        binding.cafeteriaStuKoreanMenuList.apply {
+        binding.rvCafeteriaMenuList.apply {
             val layoutManager = FlexboxLayoutManager(context)
             layoutManager.flexDirection = FlexDirection.ROW
             this.adapter = this@CafeteriaFragment.stuKoreanMenuAdapter
@@ -45,7 +41,7 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding, CafeteriaViewMo
             this.setHasFixedSize(true)
         }
 
-        binding.cafeteriaStuAnotherMenuList.apply {
+        binding.rvCafeteriaAnotherMenuList.apply {
             val layoutManager = FlexboxLayoutManager(context)
             layoutManager.flexDirection = FlexDirection.ROW
             this.adapter = this@CafeteriaFragment.stuAnotherMenuAdapter
@@ -53,35 +49,19 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding, CafeteriaViewMo
             this.setHasFixedSize(true)
         }
 
-        binding.cafeteriaEduKoreanMenuList.apply {
-            val layoutManager = FlexboxLayoutManager(context)
-            layoutManager.flexDirection = FlexDirection.ROW
-            this.adapter = this@CafeteriaFragment.eduKoreanMenuAdapter
-            this.layoutManager = layoutManager
-            this.setHasFixedSize(true)
-        }
-
-        binding.cafeteriaEduAnotherMenuList.apply {
-            val layoutManager = FlexboxLayoutManager(context)
-            layoutManager.flexDirection = FlexDirection.ROW
-            this.adapter = this@CafeteriaFragment.eduAnotherMenuAdapter
-            this.layoutManager = layoutManager
-            this.setHasFixedSize(true)
-        }
-
         val wmc =
             WindowMetricsCalculator.getOrCreate().computeCurrentWindowMetrics(requireActivity())
 
-        binding.cafeteriaCalendar.apply {
+        binding.cvCafeteriaCalendar.apply {
             val dayWidth = wmc.bounds.width() / 5
             val dayHeight: Int = (dayWidth * 1.25).toInt()
 
             daySize = Size(dayWidth, dayHeight)
         }
 
-        binding.cafeteriaCalendar.dayBinder = object : DayBinder<CafeteriaContainer> {
+        binding.cvCafeteriaCalendar.dayBinder = object : DayBinder<CafeteriaContainer> {
             override fun create(view: View): CafeteriaContainer =
-                CafeteriaContainer(view, binding.cafeteriaCalendar, viewModel)
+                CafeteriaContainer(view, binding.cvCafeteriaCalendar, viewModel)
 
             override fun bind(container: CafeteriaContainer, day: CalendarDay) = container.bind(day)
         }
@@ -98,25 +78,22 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding, CafeteriaViewMo
             showToast(getString(resId))
         })
 
-        viewModel.stuMenus.observe(viewLifecycleOwner) {
-            stuKoreanMenuAdapter.submitList(it.first)
-            stuAnotherMenuAdapter.submitList(it.second)
+        viewModel.menus.observe(viewLifecycleOwner) {
+            stuKoreanMenuAdapter.submitList(it)
+            // 일품 메뉴 : 일품 메뉴는 리스트로 제작하여 등록
+            stuAnotherMenuAdapter.submitList(listOf(getString(R.string.cafeteria_no_menu)))
         }
 
-        viewModel.eduMenus.observe(viewLifecycleOwner) {
-            eduKoreanMenuAdapter.submitList(it.first)
-            eduAnotherMenuAdapter.submitList(it.second)
-        }
     }
 
     override fun initAfterBinding() {
-        binding.cafeteriaCalendar.setup(
+        binding.cvCafeteriaCalendar.setup(
             YearMonth.now().minusMonths(2),
             YearMonth.now().plusMonths(1),
             DayOfWeek.values().random()
         )
 
-        binding.cafeteriaCalendar.scrollToDate(LocalDate.now().minusDays(2))
+        binding.cvCafeteriaCalendar.scrollToDate(LocalDate.now().minusDays(2))
 
         binding.cafeteriaErrorContainer.refresh.setOnClickListener {
             viewModel.fetchCafeteria()
@@ -127,7 +104,7 @@ class CafeteriaFragment : BaseFragment<FragmentCafeteriaBinding, CafeteriaViewMo
         super.onPause()
         notifyDateChanged(
             viewModel,
-            binding.cafeteriaCalendar,
+            binding.cvCafeteriaCalendar,
             viewModel.selectedDate.value,
             LocalDate.now()
         )
