@@ -1,9 +1,24 @@
 package com.dongyang.android.youdongknowme.data.remote.entity
 
+import com.google.gson.annotations.SerializedName
 
 data class Schedule(
-    var year: String,
-    var month: Int,
-    var date: String,
-    var content: String
+    @SerializedName("year")
+    val year: Int,
+    @SerializedName("yearSchedule")
+    val yearSchedules: List<YearSchedule>,
+)
+
+data class YearSchedule(
+    @SerializedName("month")
+    val month: Int,
+    @SerializedName("scheduleEntries")
+    val scheduleEntries: List<ScheduleEntry>
+)
+
+data class ScheduleEntry(
+    @SerializedName("content")
+    val contents: String,
+    @SerializedName("date")
+    val dates: List<String>
 )
