@@ -1,54 +1,55 @@
 package com.dongyang.android.youdongknowme.data.repository
 
-import com.dongyang.android.youdongknowme.data.local.SharedPreference
-import com.dongyang.android.youdongknowme.data.local.dao.AlarmDao
 import com.dongyang.android.youdongknowme.data.remote.entity.Notice
 import com.dongyang.android.youdongknowme.data.remote.service.NoticeService
 import com.dongyang.android.youdongknowme.standard.network.ErrorResponseHandler
 import com.dongyang.android.youdongknowme.standard.network.NetworkResult
 import com.dongyang.android.youdongknowme.standard.network.RetrofitObject
-import kotlinx.coroutines.flow.Flow
-import retrofit2.Response
 
 class NoticeRepository(
-    private val alarmDao: AlarmDao,
     private val errorResponseHandler: ErrorResponseHandler
 ) {
-    suspend fun fetchAllNotices(): NetworkResult<Map<String,List<Notice>>> {
+    suspend fun fetchUniversityNotices(page: Int): NetworkResult<List<Notice>> {
         return try {
-            val departNotices = RetrofitObject.getNetwork().create(NoticeService::class.java).getList(SharedPreference.getCode())
-            val schoolNotices = RetrofitObject.getNetwork().create(NoticeService::class.java).getList(CODE.SCHOOL_CODE)
+            val universityNotices = RetrofitObject.getNetwork().create(NoticeService::class.java)
+                .getUniversityNotice(page, DEFAULT_SIZE)
 
-            val result : Map<String,List<Notice>> = mapOf("school" to schoolNotices, "depart" to departNotices)
-
-            NetworkResult.Success(result)
+            NetworkResult.Success(universityNotices)
         } catch (exception: Exception) {
             val error = errorResponseHandler.getError(exception)
             NetworkResult.Error(error)
         }
     }
 
-    suspend fun fetchNotices(code: Int): NetworkResult<List<Notice>> {
+    suspend fun fetchDepartmentNotices(
+        department: String,
+        page: Int)
+    : NetworkResult<List<Notice>> {
         return try {
-            val response = RetrofitObject.getNetwork().create(NoticeService::class.java).getList(code)
-            NetworkResult.Success(response)
+            val departmentNotices = RetrofitObject.getNetwork().create(NoticeService::class.java)
+                .getDepartmentNotice(department, page, DEFAULT_SIZE)
+            NetworkResult.Success(departmentNotices)
         } catch (exception: Exception) {
             val error = errorResponseHandler.getError(exception)
             NetworkResult.Error(error)
         }
     }
 
-    suspend fun fetchSearchNotices(code: Int, keyword: String): NetworkResult<List<Notice>> {
+    suspend fun fetchSearchNotices(
+        searchWord: String,
+        page: Int
+    ): NetworkResult<List<Notice>> {
         return try {
-            val response = RetrofitObject.getNetwork().create(NoticeService::class.java).getSearchList(code, keyword)
-            NetworkResult.Success(response)
+            val searchNotices = RetrofitObject.getNetwork().create(NoticeService::class.java)
+                .getSearchNotice(searchWord, "컴퓨터소프트웨어공학과", page, DEFAULT_SIZE)
+            NetworkResult.Success(searchNotices)
         } catch (exception: Exception) {
             val error = errorResponseHandler.getError(exception)
             NetworkResult.Error(error)
         }
     }
 
-    fun getUnVisitedAlarmCount(): Flow<Int> = alarmDao.getUnVisitedAlarmCount()
-
-    fun getDepartmentCode(): Int = SharedPreference.getCode()
+    companion object {
+        private const val DEFAULT_SIZE = 20
+    }
 }
