@@ -6,51 +6,31 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.dongyang.android.youdongknowme.data.remote.entity.Notice
 import com.dongyang.android.youdongknowme.databinding.ItemNoticeBinding
-import com.dongyang.android.youdongknowme.ui.view.notice.NoticeClickListener
+import com.dongyang.android.youdongknowme.ui.viewholder.NoticeViewHolder
 
-class NoticeAdapter : RecyclerView.Adapter<NoticeAdapter.ViewHolder>() {
+class NoticeAdapter(private val onItemClick: (url: String) -> Unit) :
+    RecyclerView.Adapter<NoticeViewHolder>() {
 
-    init {
-        setHasStableIds(true)
-    }
-
-    private var noticeList = arrayListOf<Notice>()
-    private var itemClickListener : NoticeClickListener? = null
-
-    inner class ViewHolder(private val binding: ItemNoticeBinding)
-        : RecyclerView.ViewHolder(binding.root) {
-            fun bind(item : Notice) {
-                binding.notice = item
-                binding.itemClickListener = itemClickListener
-            }
-    }
+    private val noticeList = arrayListOf<Notice>()
 
     @SuppressLint("NotifyDataSetChanged")
-    fun submitList(item : List<Notice>) {
+    fun submitList(item: List<Notice>) {
         noticeList.clear()
         noticeList.addAll(item)
         notifyDataSetChanged()
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder(
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NoticeViewHolder {
+        return NoticeViewHolder(
             ItemNoticeBinding.inflate(
                 LayoutInflater.from(parent.context), parent, false
-            )
+            ), onItemClick
         )
     }
 
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: NoticeViewHolder, position: Int) {
         holder.bind(noticeList[position])
     }
 
-    override fun getItemId(position: Int): Long {
-        return position.toLong()
-    }
-
     override fun getItemCount(): Int = noticeList.size
-
-    fun setItemClickListener(listener : NoticeClickListener) {
-        itemClickListener = listener
-    }
 }
