@@ -1,13 +1,12 @@
 package com.dongyang.android.youdongknowme.ui.view.alarm
 
 import android.content.Intent
-import androidx.recyclerview.widget.LinearLayoutManager
+import android.view.View
 import com.dongyang.android.youdongknowme.R
 import com.dongyang.android.youdongknowme.data.local.entity.AlarmEntity
 import com.dongyang.android.youdongknowme.databinding.ActivityAlarmBinding
 import com.dongyang.android.youdongknowme.standard.base.BaseActivity
 import com.dongyang.android.youdongknowme.standard.util.mapDepartmentKoreanToCode
-import com.dongyang.android.youdongknowme.ui.adapter.AlarmAdapter
 import com.dongyang.android.youdongknowme.ui.view.detail.DetailActivity
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -15,28 +14,26 @@ class AlarmActivity : BaseActivity<ActivityAlarmBinding, AlarmViewModel>(), Alar
     override val layoutResourceId: Int = R.layout.activity_alarm
     override val viewModel: AlarmViewModel by viewModel()
 
-    private lateinit var adapter : AlarmAdapter
-
     override fun initStartView() {
-        adapter = AlarmAdapter().apply { setItemClickListener(this@AlarmActivity) }
         binding.vm = viewModel
-        binding.alarmRcv.apply {
-            this.adapter = this@AlarmActivity.adapter
-            this.layoutManager = LinearLayoutManager(this@AlarmActivity)
-            this.setHasFixedSize(true)
+        setupToolbar()
+    }
+
+    private fun setupToolbar() {
+        binding.alarmToolbar.apply {
+            tvToolbarTitle.text = getString(R.string.alarm_title)
+            btnToolbarExit.visibility = View.VISIBLE
+            tvToolbarActionButton.text = getString(R.string.alarm_action_button_text)
+            tvToolbarActionButton.visibility = View.VISIBLE
         }
     }
 
-    override fun initDataBinding() {
-        viewModel.alarmList.observe(this) {
-            adapter.submitList(it)
-        }
-    }
+    override fun initDataBinding() = Unit
 
     override fun initAfterBinding() {
         viewModel.getAlarms()
 
-        binding.alarmToolbar.toolbarExit.setOnClickListener {
+        binding.alarmToolbar.btnToolbarExit.setOnClickListener {
             finish()
         }
     }
