@@ -6,22 +6,18 @@ import com.dongyang.android.youdongknowme.R
 import com.dongyang.android.youdongknowme.data.local.entity.KeywordEntity
 import com.dongyang.android.youdongknowme.databinding.ActivityKeywordBinding
 import com.dongyang.android.youdongknowme.standard.base.BaseActivity
-import com.dongyang.android.youdongknowme.ui.view.main.MainActivity
+import com.dongyang.android.youdongknowme.ui.view.permission.OnboardingPermissionActivity
 import com.google.android.material.chip.Chip
 import com.google.android.material.chip.ChipGroup
-import com.google.firebase.FirebaseApp
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import timber.log.Timber
 
 class KeywordActivity : BaseActivity<ActivityKeywordBinding, KeywordViewModel>() {
 
     override val layoutResourceId: Int = R.layout.activity_keyword
     override val viewModel: KeywordViewModel by viewModel()
 
-    override fun initStartView() {
-        binding.vm = viewModel
-    }
-
+    override fun initStartView() = Unit
+    
     override fun initDataBinding() {
         // 효율을 위해 단 한번만 옵저빙하여 이미 구독중인 항목을 선택 처리
         viewModel.localKeywordList.observe(this, object : Observer<List<KeywordEntity>> {
@@ -29,11 +25,11 @@ class KeywordActivity : BaseActivity<ActivityKeywordBinding, KeywordViewModel>()
                 viewModel.setAllKeywords(t?.filter { it.isSubscribe }?.map { it.name }
                     ?: listOf(""))
                 setCheckChipChange(
-                    binding.keywordClassChipGroup,
-                    binding.keywordMoneyChipGroup,
-                    binding.keywordAcademicChipGroup,
-                    binding.keywordEmploymentChipGroup,
-                    binding.keywordEtcChipGroup
+                    binding.chipGroupKeywordClass,
+                    binding.chipGroupKeywordMoney,
+                    binding.chipGroupKeywordAcademic,
+                    binding.chipGroupKeywordEmployment,
+                    binding.chipGroupKeywordEtc
                 )
                 viewModel.localKeywordList.removeObserver(this)
             }
@@ -44,21 +40,18 @@ class KeywordActivity : BaseActivity<ActivityKeywordBinding, KeywordViewModel>()
         viewModel.checkFirstLaunch()
         viewModel.getLocalKeywordList()
 
-        // 뒤로가기 버튼 눌렀을 때
-        binding.keywordToolbar.toolbarExit.setOnClickListener {
-            finish()
-        }
-
         // TODO :: 안드로이드 데이터베이스에 유저별 설정한 키워드 저장 및 파이어베이스 키워드 구독 설정
-        binding.keywordCompleteBtn.setOnClickListener {
+        binding.btnKeywordComplete.setOnClickListener {
             viewModel.subscribeCheckedKeyword()
             if (viewModel.isFirstLaunch.value == true) {
                 viewModel.setFirstLaunch(false)
-                val intent = Intent(this@KeywordActivity, MainActivity::class.java)
+                val intent = Intent(this@KeywordActivity, OnboardingPermissionActivity::class.java)
                 startActivity(intent)
             }
             finish()
         }
+
+        binding.toolbarKeyword.btnToolbarExit.setOnClickListener { finish() }
     }
 
     private fun setCheckChipChange(vararg chipGroups: ChipGroup) {
