@@ -1,5 +1,7 @@
 package com.dongyang.android.youdongknowme.ui.view.main
 
+import android.content.Context
+import android.content.Intent
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.plusAssign
 import androidx.navigation.ui.setupWithNavController
@@ -7,7 +9,10 @@ import com.dongyang.android.youdongknowme.R
 import com.dongyang.android.youdongknowme.databinding.ActivityMainBinding
 import com.dongyang.android.youdongknowme.standard.base.BaseActivity
 import com.dongyang.android.youdongknowme.ui.view.util.KeepStateNavigator
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import org.koin.androidx.viewmodel.ext.android.viewModel
+import timber.log.Timber
 
 
 /* 메인 액티비티 */
@@ -22,13 +27,38 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainViewModel>() {
         val navController = navHostFragment.navController
 
         val navigator =
-            KeepStateNavigator(this, navHostFragment.childFragmentManager, binding.mainNavContainer.id)
+            KeepStateNavigator(
+                this,
+                navHostFragment.childFragmentManager,
+                binding.mainNavContainer.id
+            )
         navController.navigatorProvider += navigator
         navController.setGraph(R.navigation.dmu_navigation)
         binding.mainNvBottom.setupWithNavController(navController)
+
+        getFcmToken()
     }
 
     override fun initDataBinding() {}
 
     override fun initAfterBinding() {}
+
+    private fun getFcmToken() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                return@OnCompleteListener
+            }
+
+            val token = task.result
+            Timber.d(token)
+            //TODO: prefs 선언 중복 제거될 경우 주석 제거
+            //prefs.edit().putString("fcm_token", token).commit()
+        })
+    }
+
+    companion object {
+        fun createIntent(context: Context): Intent {
+            return Intent(context, MainActivity::class.java)
+        }
+    }
 }
