@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.dongyang.android.youdongknowme.data.local.SharedPreference.NO_SCHEDULE
 import com.dongyang.android.youdongknowme.data.remote.entity.Schedule
+import com.dongyang.android.youdongknowme.data.remote.entity.ScheduleEntry
 import com.dongyang.android.youdongknowme.data.repository.ScheduleRepository
 import com.dongyang.android.youdongknowme.standard.base.BaseViewModel
 import com.dongyang.android.youdongknowme.standard.network.NetworkResult
@@ -13,7 +14,6 @@ import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import kotlinx.coroutines.launch
-
 
 /* 학사 일정 뷰모델 */
 class ScheduleViewModel(private val scheduleRepository: ScheduleRepository) : BaseViewModel() {
@@ -27,10 +27,11 @@ class ScheduleViewModel(private val scheduleRepository: ScheduleRepository) : Ba
     private val _isError: MutableLiveData<Boolean> = MutableLiveData()
     val isError: LiveData<Boolean> = _isError
 
-    private val _scheduleList = MutableLiveData<List<Schedule>>()
-    val scheduleList: LiveData<List<Schedule>> = _scheduleList
+    private val _scheduleList = MutableLiveData<List<ScheduleEntry>>()
+    val scheduleList: LiveData<List<ScheduleEntry>> = _scheduleList
 
     private val _pickYear = MutableLiveData<Int>()
+    val pickYear: LiveData<Int> = _pickYear
 
     private val _pickMonth = MutableLiveData<Int>()
     val pickMonth: LiveData<Int> = _pickMonth
@@ -48,12 +49,16 @@ class ScheduleViewModel(private val scheduleRepository: ScheduleRepository) : Ba
                 when (val result = scheduleRepository.fetchSchedules()) {
                     is NetworkResult.Success -> {
                         val scheduleList = result.data
+
                         // 선택한 연월 조건에 따라 리스트 출력
-                        _scheduleList.postValue(scheduleList.filter { it.month == pickMonth.value && it.year == _pickYear.value.toString() })
+
+                        _scheduleList.postValue(getSchedulesForPickDate(scheduleList))
+
                         scheduleRepository.setLocalSchedules(Gson().toJson(scheduleList))
                         _isError.postValue(false)
                         _isLoading.postValue(false)
                     }
+
                     is NetworkResult.Error -> {
                         handleError(result, _errorState)
                         _isError.postValue(true)
@@ -64,17 +69,26 @@ class ScheduleViewModel(private val scheduleRepository: ScheduleRepository) : Ba
         } else {
             // 저장한 데이터가 있으면 로컬에서 곧바로 데이터를 받아와 화면에 출력
             val localSchedules = scheduleRepository.getLocalSchedules()
-            val scheduleList =
-                Gson().fromJson<List<Schedule>>(
+            val scheduleList = getSchedulesForPickDate(
+                Gson().fromJson(
                     localSchedules,
                     object : TypeToken<List<Schedule>>() {}.type
-                ).filter {
-                    // 선택한 연월 조건에 따라 리스트 출력
-                    it.month == pickMonth.value && it.year == _pickYear.value.toString()
-                }
-
+                )
+            )
             _isError.postValue(false)
             _scheduleList.postValue(scheduleList)
         }
+    }
+
+    private fun getSchedulesForPickDate(list: List<Schedule>): List<ScheduleEntry> {
+        return list
+            .filter { schedule ->
+                schedule.year == pickYear.value &&
+                        schedule.yearSchedules.any { it.month == pickMonth.value }
+            }.flatMap { schedule ->
+                schedule.yearSchedules
+                    .find { it.month == pickMonth.value }
+                    ?.scheduleEntries.orEmpty()
+            }
     }
 }
